rework core classes

+++ src/main/scala/com/rbrtbrnschn/voxel2/core/Color.scala
@@ -2,16 +2,16 @@ package com.rbrtbrnschn.voxel2.core
 
 import org.joml.Vector3f
 
-case class Color(r: Double, g: Double, b: Double) {
-  def toVec3f: Vector3f = new Vector3f(r.toFloat, g.toFloat, b.toFloat)
+case class Color(r: Float, g: Float, b: Float) {
+  def toVec3f: Vector3f = new Vector3f(r, g, b)
 }
 object Color {
-  val white: Color = new Color(1.0,1.0,1.0)
-  val black: Color = new Color(0.0,0.0,0.0)
-  val red: Color = new Color(1.0, 0.0, 0.0)
-  val green: Color = new Color(0.0, 1.0, 0.0)
-  val blue: Color = new Color(0.0, 0.0, 1.0)
-  val yellow: Color = new Color(1.0, 1.0, 0.0)
-  val cyan: Color = new Color(0.0, 1.0, 1.0)
-  val magenta: Color = new Color(1.0, 0.0, 1.0)
+  val white: Color = new Color(1.0f,1.0f,1.0f)
+  val black: Color = new Color(0.0f,0.0f,0.0f)
+  val red: Color = new Color(1.0f, 0.0f, 0.0f)
+  val green: Color = new Color(0.0f, 1.0f, 0.0f)
+  val blue: Color = new Color(0.0f, 0.0f, 1.0f)
+  val yellow: Color = new Color(1.0f, 1.0f, 0.0f)
+  val cyan: Color = new Color(0.0f, 1.0f, 1.0f)
+  val magenta: Color = new Color(1.0f, 0.0f, 1.0f)
 }
